Split sync flow to SyncStrategy
